# Performance: Use a specialized sum accumulator for retractable aggregregates